backend/requirements.txt
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
tensorflow==2.16.1
Pillow==10.2.0
flask==3.0.0
flask-cors==4.0.0
librosa==0.10.1
matplotlib==3.8.2